chore: rename `trackingModules` to `features`
