+++ lib/features/palette/PaletteTracking.js
