+++ lib/features/palette/index.js
