+++ lib/index.js
@@ -1,2 +1,2 @@
 export { default as BpmnJSTracking } from './BpmnJSTracking';
-export { default as BpmnJSTrackingModules } from './trackingModules';
+export { default as BpmnJSTrackingModules } from './features';
